OE0-37: Moved PatientCategoryMapping to patientMapping.py
