+++ api_fhir_r4/mapping/__init__.py
@@ -1,6 +1,6 @@
 from api_fhir_r4.mapping import contractMapping
 from api_fhir_r4.mapping import patientMapping
-from api_fhir_r4.mapping import actvityDefinitionMapping
+from api_fhir_r4.mapping import activityDefinitionMapping
 from api_fhir_r4.mapping import locationMapping
 from api_fhir_r4.mapping import groupMapping
 from api_fhir_r4.mapping import medicationMapping
